third_party/WebKit/LayoutTests/fast/encoding/api/byte-order-marks-expected.txt

Test the Encoding API's handling of byte-order marks (BOMs).

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS new TextDecoder('utf-8').decode(new Uint8Array(utf8)) is "zÂ¢æ°´ğ„ô¿½"
PASS new TextDecoder('utf-16le').decode(new Uint8Array(utf16le)) is "zÂ¢æ°´ğ„ô¿½"
PASS new TextDecoder('utf-16be').decode(new Uint8Array(utf16be)) is "zÂ¢æ°´ğ„ô¿½"
PASS new TextDecoder('utf-8').decode(new Uint8Array(utf8_bom.concat(utf8))) is "zÂ¢æ°´ğ„ô¿½"
PASS new TextDecoder('utf-16le').decode(new Uint8Array(utf16le_bom.concat(utf16le))) is "zÂ¢æ°´ğ„ô¿½"
PASS new TextDecoder('utf-16be').decode(new Uint8Array(utf16be_bom.concat(utf16be))) is "zÂ¢æ°´ğ„ô¿½"
PASS new TextDecoder('utf-8').decode(new Uint8Array(utf16le_bom.concat(utf8))) is not "zÂ¢æ°´ğ„ô¿½"
PASS new TextDecoder('utf-8').decode(new Uint8Array(utf16be_bom.concat(utf8))) is not "zÂ¢æ°´ğ„ô¿½"
PASS new TextDecoder('utf-16le').decode(new Uint8Array(utf8_bom.concat(utf16le))) is not "zÂ¢æ°´ğ„ô¿½"
PASS new TextDecoder('utf-16le').decode(new Uint8Array(utf16be_bom.concat(utf16le))) is not "zÂ¢æ°´ğ„ô¿½"
PASS new TextDecoder('utf-16be').decode(new Uint8Array(utf8_bom.concat(utf16be))) is not "zÂ¢æ°´ğ„ô¿½"
PASS new TextDecoder('utf-16be').decode(new Uint8Array(utf16le_bom.concat(utf16be))) is not "zÂ¢æ°´ğ„ô¿½"
PASS successfullyParsed is true

TEST COMPLETE
